Add ErrorDisplayable to factorize error handling
